+++ api/src/main/java/org/apache/cloudstack/api/command/user/volume/ListVolumesCmd.java
@@ -88,7 +88,7 @@ public class ListVolumesCmd extends BaseListTaggedResourcesCmd implements UserCm
             RoleType.Admin})
     private Boolean display;
 
-    @Parameter(name = ApiConstants.LIST_SYSTEM_VMS, type = CommandType.BOOLEAN, description = "list system VMs; only ROOT admin is eligible to pass this parameter", since = "4.17",
+    @Parameter(name = ApiConstants.LIST_SYSTEM_VMS, type = CommandType.BOOLEAN, description = "list system VMs; only ROOT admin is eligible to pass this parameter", since = "4.18",
             authorized = { RoleType.Admin })
     private Boolean listSystemVms;
